split mainTheme.ts into more sizeable files
