+++ src/pages/registration/resource_type_tab/sub_type_forms/artistic_types/moving_picture/ArtisticMovingPictureForm.tsx
@@ -16,7 +16,7 @@ import { ErrorMessage, Field, FieldArray, FieldArrayRenderProps, FieldProps, use
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { StyledSelectWrapper } from '../../../../../../components/styled/Wrappers';
-import { alternatingTableRowColor } from '../../../../../../themes/mainTheme';
+import { alternatingTableRowColor } from '../../../../../../themes/cssStyles';
 import {
   ArtisticRegistration,
   FilmOutput,
