+++ src/themes/colors.ts
@@ -1,0 +1,28 @@
+// Colors: https://www.figma.com/file/3hggk6SX2ca81U8kwaZKFs/Farger-NVA
+export enum Color {
+  Black = '#222222',
+  White = '#FFFFFF',
+  Purple = '#7351FB',
+  PurpleLight = '#D1CDFF',
+  PurpleDark = '#0A0132',
+  TextPrimary = 'rgba(0, 0, 0, 0.87)',
+  CentralImportMain = '#D9D9D9',
+  Registration = '#EED2AE',
+  Person = '#BAD2F7',
+  Project = '#E8B8D0',
+  Neutral97 = '#F7F7F7',
+  Neutral95 = '#F1F1F1',
+  Neutral87 = '#DEDEDE',
+  Neutral46 = '#767676',
+  BlueLight = '#E6F0FF',
+  BlueClear = '#75C7F0',
+  Blue = '#004FCF',
+  Green = '#096638',
+  GreenLight = '#CFF7E2',
+  Red = '#B60203',
+  RedLight = '#FFEAE9',
+  YellowLight = '#FCEED2',
+  YellowClear = '#FBE774',
+  Yellow = '#FFB700',
+  Orange = '#FFAC70',
+}
